requirements.txt
Flask
setuptools
pytest
SQLAlchemy
psycopg2
pyjwt
pykafka